requirements.txt
autopep8==1.6.0
lxml==4.7.1
pikepdf==4.2.0
Pillow==8.4.0
pycodestyle==2.8.0
toml==0.10.2
